requirements.txt
bcrypt==4.0.1
greenlet==2.0.2
PyMySQL==1.0.2
python-dotenv==1.0.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
